+++ data/pnw.jl
@@ -1,0 +1,158 @@
+# from https://github.com/jakelawlor/PNWColors
+# license unspecified but apparently CreativeCommons0
+
+loadcolorscheme(:Starfish, [
+   RGB(0.141,0.286,0.18),
+   RGB(0.004,0.357,0.345),
+   RGB(0.173,0.38,0.518),
+   RGB(0.349,0.384,0.608),
+   RGB(0.537,0.408,0.616),
+   RGB(0.729,0.475,0.6),
+   RGB(0.902,0.608,0.6),
+        ],
+    "PNW Pacific North West",
+    "Washington State, Pacific Northwest")
+
+loadcolorscheme(:Shuksan, [
+   RGB(0.2,0.153,0.118),
+   RGB(0.455,0.404,0.494),
+   RGB(0.675,0.557,0.671),
+   RGB(0.843,0.694,0.773),
+   RGB(0.922,0.741,0.784),
+   RGB(0.949,0.808,0.78),
+   RGB(0.973,0.89,0.82),
+   RGB(0.996,0.984,0.914),
+        ],
+    "PNW Pacific North West",
+    "Washington State, Pacific Northwest")
+
+loadcolorscheme(:Bay, [
+   RGB(0.0,0.286,0.435),
+   RGB(0.059,0.522,0.627),
+   RGB(0.929,0.843,0.275),
+   RGB(0.929,0.545,0.0),
+   RGB(0.867,0.255,0.141),
+        ],
+    "PNW Pacific North West",
+    "Washington State, Pacific Northwest")
+
+loadcolorscheme(:Winter, [
+   RGB(0.176,0.161,0.149),
+   RGB(0.2,0.271,0.306),
+   RGB(0.325,0.451,0.502),
+   RGB(0.506,0.663,0.678),
+   RGB(0.925,0.925,0.925),
+        ],
+    "PNW Pacific North West",
+    "Washington State, Pacific Northwest")
+
+loadcolorscheme(:Lake, [
+   RGB(0.212,0.161,0.016),
+   RGB(0.329,0.271,0.059),
+   RGB(0.271,0.408,0.118),
+   RGB(0.29,0.569,0.322),
+   RGB(0.392,0.659,0.659),
+   RGB(0.522,0.714,0.808),
+   RGB(0.804,0.898,0.976),
+   RGB(0.933,0.953,1.0),
+        ],
+    "PNW Pacific North West",
+    "Washington State, Pacific Northwest")
+
+loadcolorscheme(:Sunset, [
+   RGB(0.255,0.278,0.42),
+   RGB(0.404,0.329,0.471),
+   RGB(0.62,0.388,0.455),
+   RGB(0.776,0.482,0.435),
+   RGB(0.871,0.608,0.443),
+   RGB(0.937,0.737,0.51),
+   RGB(0.984,0.875,0.635),
+        ],
+    "PNW Pacific North West",
+    "Washington State, Pacific Northwest")
+
+loadcolorscheme(:Shuksan2, [
+   RGB(0.365,0.455,0.647),
+   RGB(0.69,0.796,0.906),
+   RGB(0.996,0.969,0.78),
+   RGB(0.922,0.627,0.494),
+   RGB(0.659,0.333,0.306),
+        ],
+    "PNW Pacific North West",
+    "Washington State, Pacific Northwest")
+
+loadcolorscheme(:Cascades, [
+   RGB(0.176,0.251,0.188),
+   RGB(0.318,0.408,0.137),
+   RGB(0.871,0.753,0.0),
+   RGB(0.886,0.886,0.376),
+   RGB(0.404,0.494,0.557),
+   RGB(0.533,0.635,0.725),
+        ],
+    "PNW Pacific North West",
+    "Washington State, Pacific Northwest")
+
+loadcolorscheme(:Sailboat, [
+   RGB(0.431,0.486,0.725),
+   RGB(0.482,0.737,0.835),
+   RGB(0.816,0.886,0.686),
+   RGB(0.961,0.859,0.6),
+   RGB(0.91,0.612,0.506),
+   RGB(0.824,0.518,0.553),
+        ],
+    "PNW Pacific North West",
+    "Washington State, Pacific Northwest")
+
+loadcolorscheme(:Moth, [
+   RGB(0.29,0.227,0.231),
+   RGB(0.596,0.255,0.212),
+   RGB(0.761,0.416,0.478),
+   RGB(0.925,0.753,0.631),
+   RGB(0.941,0.941,0.894),
+        ],
+    "PNW Pacific North West",
+    "Washington State, Pacific Northwest")
+
+loadcolorscheme(:Spring, [
+   RGB(0.847,0.682,0.867),
+   RGB(0.749,0.608,0.867),
+   RGB(0.796,0.455,0.678),
+   RGB(0.902,0.62,0.612),
+   RGB(1.0,0.765,0.639),
+   RGB(0.984,0.894,0.776),
+        ],
+    "PNW Pacific North West",
+    "Washington State, Pacific Northwest")
+
+loadcolorscheme(:Mushroom, [
+   RGB(0.31,0.255,0.169),
+   RGB(0.525,0.353,0.235),
+   RGB(0.729,0.471,0.243),
+   RGB(0.902,0.612,0.298),
+   RGB(0.984,0.8,0.455),
+   RGB(1.0,0.984,0.855),
+        ],
+    "PNW Pacific North West",
+    "Washington State, Pacific Northwest")
+
+loadcolorscheme(:Sunset2, [
+   RGB(0.114,0.271,0.498),
+   RGB(0.38,0.349,0.616),
+   RGB(0.765,0.388,0.467),
+   RGB(0.922,0.498,0.329),
+   RGB(0.949,0.686,0.29),
+        ],
+    "PNW Pacific North West",
+    "Washington State, Pacific Northwest")
+
+loadcolorscheme(:Anemone, [
+   RGB(0.0,0.58,0.455),
+   RGB(0.067,0.761,0.71),
+   RGB(0.447,0.882,0.882),
+   RGB(0.945,0.957,0.933),
+   RGB(0.937,0.867,0.812),
+   RGB(0.863,0.745,0.608),
+   RGB(0.69,0.596,0.424),
+        ],
+    "PNW Pacific North West",
+    "Washington State, Pacific Northwest")