+++ data/wesanderson.jl
@@ -1,0 +1,245 @@
+# https://github.com/karthik/wesanderson
+# unknown license 
+
+loadcolorscheme(:BottleRocket1, [
+        RGB{N0f8}(0.643, 0.157, 0.125),
+        RGB{N0f8}(0.373, 0.337, 0.278),
+        RGB{N0f8}(0.608, 0.067, 0.055),
+        RGB{N0f8}(0.247, 0.318, 0.318),
+        RGB{N0f8}(0.306, 0.165, 0.118),
+        RGB{N0f8}(0.333, 0.012, 0.027),
+        RGB{N0f8}(0.047, 0.09, 0.027),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:BottleRocket2, [
+        RGB{N0f8}(0.98, 0.835, 0.063),
+        RGB{N0f8}(0.796, 0.137, 0.078),
+        RGB{N0f8}(0.153, 0.188, 0.275),
+        RGB{N0f8}(0.208, 0.282, 0.137),
+        RGB{N0f8}(0.118, 0.118, 0.118),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:Rushmore1, [
+        RGB{N0f8}(0.882, 0.741, 0.427),
+        RGB{N0f8}(0.918, 0.745, 0.58),
+        RGB{N0f8}(0.043, 0.467, 0.369),
+        RGB{N0f8}(0.208, 0.153, 0.29),
+        RGB{N0f8}(0.949, 0.188, 0.059),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:Rushmore, [
+        RGB{N0f8}(0.882, 0.741, 0.427),
+        RGB{N0f8}(0.918, 0.745, 0.58),
+        RGB{N0f8}(0.043, 0.467, 0.369),
+        RGB{N0f8}(0.208, 0.153, 0.29),
+        RGB{N0f8}(0.949, 0.188, 0.059),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:Royal1, [
+        RGB{N0f8}(0.537, 0.616, 0.643),
+        RGB{N0f8}(0.788, 0.2, 0.071),
+        RGB{N0f8}(0.98, 0.937, 0.82),
+        RGB{N0f8}(0.863, 0.525, 0.231),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:Royal2, [
+        RGB{N0f8}(0.604, 0.533, 0.133),
+        RGB{N0f8}(0.961, 0.804, 0.706),
+        RGB{N0f8}(0.973, 0.686, 0.659),
+        RGB{N0f8}(0.992, 0.867, 0.627),
+        RGB{N0f8}(0.455, 0.627, 0.537),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:Zissou1, [
+        RGB{N0f8}(0.231, 0.604, 0.698),
+        RGB{N0f8}(0.471, 0.718, 0.773),
+        RGB{N0f8}(0.922, 0.8, 0.165),
+        RGB{N0f8}(0.882, 0.686, 0.0),
+        RGB{N0f8}(0.949, 0.102, 0.0),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:Zissou1Continuous, [
+        RGB{N0f8}(0.227, 0.604, 0.698),
+        RGB{N0f8}(0.435, 0.698, 0.757),
+        RGB{N0f8}(0.569, 0.729, 0.714),
+        RGB{N0f8}(0.647, 0.761, 0.639),
+        RGB{N0f8}(0.741, 0.784, 0.506),
+        RGB{N0f8}(0.863, 0.796, 0.306),
+        RGB{N0f8}(0.89, 0.718, 0.063),
+        RGB{N0f8}(0.906, 0.596, 0.02),
+        RGB{N0f8}(0.925, 0.478, 0.02),
+        RGB{N0f8}(0.937, 0.341, 0.012),
+        RGB{N0f8}(0.945, 0.106, 0.0),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:Darjeeling1, [
+        RGB{N0f8}(1.0, 0.0, 0.0),
+        RGB{N0f8}(0.0, 0.627, 0.541),
+        RGB{N0f8}(0.949, 0.678, 0.0),
+        RGB{N0f8}(0.976, 0.518, 0.0),
+        RGB{N0f8}(0.357, 0.737, 0.839),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:Darjeeling2, [
+        RGB{N0f8}(0.925, 0.796, 0.682),
+        RGB{N0f8}(0.016, 0.424, 0.604),
+        RGB{N0f8}(0.839, 0.612, 0.306),
+        RGB{N0f8}(0.671, 0.867, 0.871),
+        RGB{N0f8}(0.0, 0.0, 0.0),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:Chevalier1, [
+        RGB{N0f8}(0.267, 0.392, 0.333),
+        RGB{N0f8}(0.992, 0.824, 0.384),
+        RGB{N0f8}(0.827, 0.867, 0.863),
+        RGB{N0f8}(0.78, 0.694, 0.612),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:FantasticFox1, [
+        RGB{N0f8}(0.867, 0.553, 0.161),
+        RGB{N0f8}(0.886, 0.824, 0.0),
+        RGB{N0f8}(0.275, 0.675, 0.784),
+        RGB{N0f8}(0.898, 0.525, 0.004),
+        RGB{N0f8}(0.706, 0.059, 0.125),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:Moonrise1, [
+        RGB{N0f8}(0.953, 0.875, 0.424),
+        RGB{N0f8}(0.808, 0.671, 0.027),
+        RGB{N0f8}(0.835, 0.835, 0.827),
+        RGB{N0f8}(0.141, 0.157, 0.102),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:Moonrise2, [
+        RGB{N0f8}(0.475, 0.557, 0.529),
+        RGB{N0f8}(0.761, 0.49, 0.22),
+        RGB{N0f8}(0.8, 0.773, 0.569),
+        RGB{N0f8}(0.161, 0.129, 0.122),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:Moonrise3, [
+        RGB{N0f8}(0.522, 0.831, 0.89),
+        RGB{N0f8}(0.957, 0.71, 0.741),
+        RGB{N0f8}(0.612, 0.588, 0.29),
+        RGB{N0f8}(0.804, 0.753, 0.549),
+        RGB{N0f8}(0.98, 0.843, 0.482),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:Cavalcanti1, [
+        RGB{N0f8}(0.847, 0.718, 0.039),
+        RGB{N0f8}(0.008, 0.251, 0.106),
+        RGB{N0f8}(0.635, 0.643, 0.459),
+        RGB{N0f8}(0.506, 0.659, 0.553),
+        RGB{N0f8}(0.592, 0.176, 0.082),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:GrandBudapest1, [
+        RGB{N0f8}(0.945, 0.733, 0.482),
+        RGB{N0f8}(0.992, 0.392, 0.404),
+        RGB{N0f8}(0.357, 0.102, 0.094),
+        RGB{N0f8}(0.839, 0.447, 0.212),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:GrandBudapest2, [
+        RGB{N0f8}(0.902, 0.627, 0.769),
+        RGB{N0f8}(0.776, 0.804, 0.969),
+        RGB{N0f8}(0.847, 0.643, 0.6),
+        RGB{N0f8}(0.447, 0.58, 0.831),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:IsleofDogs1, [
+        RGB{N0f8}(0.6, 0.525, 0.647),
+        RGB{N0f8}(0.475, 0.251, 0.18),
+        RGB{N0f8}(0.8, 0.729, 0.447),
+        RGB{N0f8}(0.059, 0.051, 0.055),
+        RGB{N0f8}(0.851, 0.816, 0.827),
+        RGB{N0f8}(0.553, 0.525, 0.502),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:IsleofDogs2, [
+        RGB{N0f8}(0.918, 0.827, 0.749),
+        RGB{N0f8}(0.667, 0.58, 0.525),
+        RGB{N0f8}(0.714, 0.522, 0.302),
+        RGB{N0f8}(0.224, 0.192, 0.184),
+        RGB{N0f8}(0.11, 0.09, 0.094),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:FrenchDispatch, [
+        RGB{N0f8}(0.565, 0.831, 0.8),
+        RGB{N0f8}(0.741, 0.188, 0.153),
+        RGB{N0f8}(0.69, 0.686, 0.635),
+        RGB{N0f8}(0.498, 0.753, 0.776),
+        RGB{N0f8}(0.616, 0.612, 0.522),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:AsteroidCity1, [
+        RGB{N0f8}(0.039, 0.624, 0.616),
+        RGB{N0f8}(0.808, 0.694, 0.459),
+        RGB{N0f8}(0.898, 0.306, 0.129),
+        RGB{N0f8}(0.424, 0.525, 0.271),
+        RGB{N0f8}(0.757, 0.529, 0.282),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:AsteroidCity2, [
+        RGB{N0f8}(0.773, 0.18, 0.098),
+        RGB{N0f8}(0.675, 0.592, 0.396),
+        RGB{N0f8}(0.329, 0.847, 0.694),
+        RGB{N0f8}(0.714, 0.486, 0.231),
+        RGB{N0f8}(0.09, 0.318, 0.286),
+        RGB{N0f8}(0.686, 0.306, 0.141),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")
+
+loadcolorscheme(:AsteroidCity3, [
+        RGB{N0f8}(0.984, 0.655, 0.165),
+        RGB{N0f8}(0.827, 0.831, 0.847),
+        RGB{N0f8}(0.796, 0.478, 0.361),
+        RGB{N0f8}(0.341, 0.522, 0.757),
+    ],
+    "Wes Anderson",
+    "Wes Anderson")