+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Added
 
+- Wes Anderson palettes (https://github.com/karthik/wesanderson)
+
+- Pacific North West palettes (https://github.com/jakelawlor/PNWColors)
+
 ## Changed
 
 - updated ScientificColorSchemes to version 8.0.0. Note that some schemes (grayC bilbao lajolla acton bamako batlowK tokyo) have changed since their 7.0 release - see their version history for details.